Fix POS month reports and checkout, Sheets payment staff names, and louder alerts.
Staff PWA push, Telegram operations webhook, daily summary, and related Hostinger schema stay in the same ship.

+++ pms_core/api_endpoints/guest_create_razorpay_order.php
@@ -42,7 +42,7 @@ if (!$rz) {
 }
 
 $receipt = 'gst_' . $bookingId . '_' . time();
-$result = $rz->createOrder(round($amount * 100), 'INR', $receipt, [
+$result = $rz->createOrder((int)round($amount * 100), 'INR', $receipt, [
     'booking_id' => (string)$bookingId,
     'property_id' => (string)$propertyId,
 ]);
